% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendesign.R
\name{designDF}
\alias{designDF}
\title{Extract design}
\usage{
designDF(x, selection)
}
\arguments{
\item{x}{a \code{DEdesign} object (output from \code{gendesign}).}

\item{selection}{choose which design to output: \code{Design} based on input or \code{suggestedDesign}. Default is \code{Design}.}
}
\value{
a data frame giving flowcell, lane and adapter assignment for each experimental unit
}
\description{
Extracts desgins from a \code{DEdesign} object
}
\examples{
treatments <- data.frame(trt = letters[1:6], replicates = rep(3,6))
des <- gendesign(treatments=treatments, nperlane=4, search.surrounding = 2)
designDF(des,selection="Design")
designDF(des,selection="suggestedDesign")

}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
