% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onedesign}
\alias{.onedesign}
\title{Calculate design for a specified nperlane}
\usage{
.onedesign(treatments, nperlane, seed, searches)
}
\arguments{
\item{treatments}{reformated input treatments dataframe}

\item{nperlane}{An integer for number of samples per lane for sequencing. Default is 4. This is used to generate \code{Design} in results.}

\item{seed}{an integer initializing the random number generator. The default is \code{seed=1}. Designs can be rebuilt repeatedly using different seed to check that a near-optimum design has been found.}

\item{searches}{the maximum number of local optima searched at each stage of a treatment and block design optimization. The default depends on the design size. For optimum results, try large number of searches.}
}
\value{
blocksdesign output for one nperlane value
}
\description{
Calculate design for a specified nperlane
}
