% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reformat}
\alias{.reformat}
\title{reformat input treatments data.frame}
\usage{
.reformat(treatments)
}
\arguments{
\item{treatments}{a \code{data.frame} showing number of samples per treatment group. For single factor experiment, \code{treatments} should contain two columns with one column showing levels of the factor and the other column named \code{"replicates"} for number of samples in each level. For factorial experiments, there should be one column for each factor, and the \code{"replicates"} column gives number of samples in each combination of factor levels. Please see \code{treatments.example} for a template. For the purpose of plotting, please use short factor names and in factorial design using numbers for factor levels is recommended.}
}
\value{
a data frame of reformated input treatments dataframe to use as blocksdesign input, repeat each row "replicates" times, so nrow = total number of samples
}
\description{
reformat input treatments data.frame
}
