% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.whichHighestEffi}
\alias{.whichHighestEffi}
\title{Find nperlane for suggestedDesign}
\usage{
.whichHighestEffi(cachedRes, target = "lane")
}
\arguments{
\item{cachedRes}{a list data frame with columns "nperlane_torun" and "res_list" (design output based on nperlane given)}

\item{target}{which block efficiency should be used to select suggestedDesign, default is "lane"}
}
\value{
\code{nperlane} value that gives the suggestedDesign
}
\description{
Find nperlane for suggestedDesign
}
