% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendesign.R
\name{efficiency}
\alias{efficiency}
\title{Extract efficiency}
\usage{
efficiency(x, selection)
}
\arguments{
\item{x}{a \code{DEdesign} object (output from \code{gendesign}).}

\item{selection}{choose which design to output: "\code{Design}" based on input or "\code{suggestedDesign}". Default is "\code{Design}".}
}
\value{
a data frame giving D-efficiencies of flowcell (If there are more than one flowcell), lane and adapter blocks
}
\description{
Extracts D-efficiency from a \code{DEdesign} object
}
\examples{
des <- gendesign(treatments=letters[1:6], replicates=rep(3,6), nperlane=4)
efficiency(des,selection="Design")
efficiency(des,selection="suggestedDesign")

}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
