% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendesign.R
\name{gendesign}
\alias{gendesign}
\title{Statistical design for RNA-seq experiments aimed at identifying differentially expressed genes}
\usage{
gendesign(treatments, replicates, nperlane = NULL, seed = 1)
}
\arguments{
\item{treatments}{a character vector for treatment groups. If the experiment consists of two or more factors, provide a character vector representing all possible combinations of levels across these factors.}

\item{replicates}{an integer vector for number of replicates for each element of \code{treatments}. Length of \code{replicates} has to match length of \code{treatments}.}

\item{nperlane}{An integer for number of samples per lane for sequencing. When \code{nperlane} is not provided, a number between 4-6 will be selected based on number of treatments to generate efficient design}

\item{seed}{an integer initializing the random number generator. The default is \code{seed=1}}
}
\value{
\code{input} A list showing input parameters to the function

\code{Design} A list with two elements: \code{design} and \code{BlocksEfficiency}. \code{design} is a data frame giving lane and adaptor assignment for treatment groups. \code{BlocksEfficiency} is a data frame giving block efficiencies (D-Efficiencies) for lane and adaptor.

\code{suggestedDesign} When possible, \code{suggestedDesign} uses different number of samples per lane from the input to give more efficient block design; otherwise it is the same as \code{Design}.  \code{suggestedDesign} is a list with two elements: \code{design} and \code{BlocksEfficiency}. \code{design} is a data frame giving lane and adaptor assignment for treatment groups. \code{BlocksEfficiency} is a data frame giving block efficiencies (D-Efficiencies) for lane and adaptor.
}
\description{
Generates block design to assign samples to sequencing lanes and adaptors on an Illumina flow cell
}
\examples{
gendesign(treatments=letters[1:4], replicates=rep(4,4), nperlane=NULL)
gendesign(treatments=letters[1:4], replicates=c(2,4,4,2), nperlane=3)
gendesign(treatments=letters[1:6], replicates=rep(3,6), nperlane=4)

}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
