% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.design.R
\name{plot.design}
\alias{plot.design}
\title{Field map of RNA-seq experimental design}
\usage{
\method{plot}{design}(designtable)
}
\arguments{
\item{designtable}{a data frame for lane and adaptor assignment. Use output from \code{gen.design}.It should have 4 columns with column names \code{chip, lane, adp, trt}.}
}
\value{
None
}
\description{
Plot the RNA-seq experimental design in the format of Illumina 8-lane flow cell
}
\examples{
des <- gen.design(trts=letters[1:6], reps=rep(3,6), nperlane=4)
plot.design (des$Design$design)
plot.design (des$suggestedDesign$design)

}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
