% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdesign.R
\name{plotdesign}
\alias{plotdesign}
\title{Field map of RNA-seq experimental design}
\usage{
plotdesign(designtable)
}
\arguments{
\item{designtable}{a data frame for lane and adaptor assignment. Use output from \code{gen.design}.It should have 4 columns with column names \code{flowcell, lane, adaptor, treatment}.}
}
\value{
None
}
\description{
Plot the RNA-seq experimental design in the format of Illumina 8-lane flow cell
}
\examples{
des <- gendesign(treatments=letters[1:6], replicates=rep(3,6), nperlane=4)
plotdesign (des$Design$design)
plotdesign (des$suggestedDesign$design)

}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
