% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendesign.R
\name{gendesign}
\alias{gendesign}
\title{Statistical design for RNA-seq experiments aimed at identifying differentially expressed genes}
\usage{
gendesign(treatments, nperlane = 4, search.surrounding = 0, seed = 1,
  searches = NULL)
}
\arguments{
\item{treatments}{a \code{data.frame} showing number of samples per treatment group. For single factor experiment, \code{treatments} should contain two columns with one column showing levels of the factor and the other column named \code{"replicates"} for number of samples in each level. For factorial experiments, there should be one column for each factor, and the \code{"replicates"} column gives number of samples in each combination of factor levels. Please see examples for a template. For the purpose of plotting, please use short factor names and in factorial design using numbers for factor levels is recommended.}

\item{nperlane}{An integer for number of samples per lane for sequencing. Default is 4. This is used to generate \code{Design} in results.}

\item{search.surrounding}{An non-negative integer with default 0, which means to only find design for number of samples per lane defined by \code{nperlane}. For other values, designs using \code{min(3,(nperlane-search.surrounding))} to \code{(nperlane+search.surrounding)} samples per lane will be compared to give \code{suggestedDesign}. \code{suggestedDesign} is the design with minimal number of flowcells among all candidate designs that gives the highest lane block efficiency (when more than one designs meet these criteria, design with highest nperlane is selected).}

\item{seed}{an integer initializing the random number generator. The default is \code{seed=1}. Designs can be rebuilt repeatedly using different seed to check that a near-optimum design has been found.}

\item{searches}{the maximum number of local optima searched at each stage of a treatment and block design optimization. The default depends on the design size. For optimum results, try large number of searches.}
}
\value{
\code{gendesign} returns an object of class "\code{DEdesign}", which is a list containing the following components:

\code{input} A list showing input parameters to the function

\code{Design} A list with two elements: \code{design} and \code{BlocksEfficiency}. \code{design} is a data frame giving flowcell, lane and adapter assignment for treatment groups. \code{BlocksEfficiency} is a data frame giving block efficiencies (D-Efficiencies) for lane and adapter (and/or flowcell when applicable).

\code{suggestedDesign} When \code{search.surrounding >0}, a list with two elements are returned here: \code{design} and \code{BlocksEfficiency}. \code{design} is a data frame giving flowcell, lane and adapter assignment for treatment groups. \code{BlocksEfficiency} is a data frame giving block efficiencies (D-Efficiencies) for lane and adapter (and/or flowcell when applicable). Designs using \code{min(3,(nperlane-search.surrounding))} to \code{(nperlane+search.surrounding)} samples per lane will be compared to give \code{suggestedDesign}. \code{suggestedDesign} is the design with minimal number of flowcells among all candidate designs that gives the highest lane block efficiency (when more than one designs meet these criteria, design with highest nperlane is selected).
}
\description{
Generates block design to assign samples to sequencing lanes and adapters for Illumina flow cell. The \code{gendesign} function is the main function of the \code{DEdesign} package. It calls the \code{\link[blocksdesign:design]{design}} function from \code{blocksdesign} package (by Dr. Rodney Edmondson), which can construct nested and crossed block designs for factorial treatments.
}
\examples{
gendesign(treatments= data.frame(trt = letters[1:4], replicates = rep(4,4)))

treatments <- data.frame(expand.grid(A=factor(1:2), B=factor(1:5)),replicates = 2)
des <- gendesign(treatments=treatments, nperlane=4, search.surrounding = 2)
designDF(des,selection="Design")
designDF(des,selection="suggestedDesign")
efficiency(des,selection="Design")
efficiency(des,selection="suggestedDesign")


}
\keyword{Illumina}
\keyword{RNA-seq,}
\keyword{block}
\keyword{cell}
\keyword{design,}
\keyword{experimental}
\keyword{flow}
\keyword{statistical}
